changes to how the map gets the parcel layer (#1276)
